scripts/requirements.txt
numpy
Pillow
scipy
torchvision>=0.15.0
transformers
diffusers
accelerate
tqdm